+++ backend/backend/settings/prod.py
@@ -4,6 +4,7 @@ from .deps.cors import *
 from .deps.org import *
 from .deps.restframework import *
 
+from substrapp.storages.minio import MinioStorage
 
 DEBUG = False
 
@@ -32,6 +33,12 @@ DATABASES = {
         'PORT': 5432,
     }
 }
+
+DATASAMPLE_BUCKET_NAME = "substra-datasample"
+DATASAMPLE_STORAGE = MinioStorage(DATASAMPLE_BUCKET_NAME)
+
+MODEL_BUCKET_NAME = "substra-model"
+MODEL_STORAGE = MinioStorage(MODEL_BUCKET_NAME)
 
 MEDIA_ROOT = os.environ.get('MEDIA_ROOT', f'/substra/medias/{ORG_NAME}')